two_factor_authentication/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
pyotp==2.9.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.1

# psycopg2-binary==2.9.9

